Support branches in git downloader.

+++ allmychanges/downloader.py
@@ -100,19 +100,37 @@ def fake_downloader(source):
         return destination
 
 
+def split_branch(url):
+    if '@' in url:
+        return url.split('@', 1)
+    return url, None
+
+
 def git_downloader(source):
     path = tempfile.mkdtemp(dir=settings.TEMP_DIR)
     url, username, repo_name = normalize_url(source)
 
+    url, branch = split_branch(url)
+
     with cd(path):
         response = envoy.run('git clone {url} {path}'.format(url=url,
                                                              path=path))
-    if response.status_code != 0:
-        if os.path.exists(path):
-            shutil.rmtree(path)
-        raise RuntimeError('Bad status_code from git clone: {0}. '
-                           'Git\'s stderr: {1}'.format(
-                               response.status_code, response.std_err))
+        if response.status_code != 0:
+            if os.path.exists(path):
+                shutil.rmtree(path)
+            raise RuntimeError('Bad status_code from git clone: {0}. '
+                               'Git\'s stderr: {1}'.format(
+                                   response.status_code, response.std_err))
+
+        if branch:
+            response = envoy.run('git checkout -b {branch} origin/{branch}'.format(branch=branch))
+
+            if response.status_code != 0:
+                if os.path.exists(path):
+                    shutil.rmtree(path)
+                raise RuntimeError('Bad status_code from git checkout -b {0}: {1}. '
+                                   'Git\'s stderr: {2}'.format(
+                                       branch, response.status_code, response.std_err))
 
     return path
